The monitoring-bot now concatenates subsections into the main road section.
The problem with the subsections is that they are not persistent.
They change in number and distance almost every few seconds.
The result is not comparable data and one would have to manually merge them.

The bot now can be configured to do this for you, calcuating the average speed over
the persistent section.